Give feeds on blogs with the plugin inactive a short TTL (#366)
* Give feeds on blogs with the plugin inactive a short TTL

Some servers have several networks running on one install and the server
admin may forget to activate the plugin on each network. Or a network
admin may disable the plugin for some reason on one blog, but caching
will still occur because it's loaded by WordPress because the WP_CACHE
constant is defined.
Unfortunately this stops wp-cache.php loading where wp_cache_setting()
is defined. That function is used to write a timestamp to the config
file whenever a post is updated. When the plugin is inactive, the feeds
will be given a TTL of 60 seconds so they do not become stale.
ref:
https://wordpress.org/support/topic/fatal-error-while-updating-post-or-publishing-new-one/

* If function DOESN'T exist, then short ttl

Reverse this logic. If the function doesn't exist then the plugin is not
activated and then set the short TTL.