Sync: Force send options, network_options, functions, constans and the user that is making the connection
On initial sync. Also make sure that we don't send all the constants and callabls on the connection.
This insures that when the user first connects the shadow site is populatd with the most relevent info and that
we don't send duplicate infomation.

This will hopefully fix bugs where the initail shadow site is left in a bad state. Since before this we were not
sending the information as soon as possibl and also we were missing all the jetpack options that we cared about.