Onboarding: create the token as a network option.
Even though this is valid for all sites in the network, it doesn't pose a threat because the user_email must match a user added to the site and the WP REST API endpoints will provide the proper validation for that. The onboarding token is just one part of the validation to ensure that the request comes from a trusted site but the REST validation will check that the user has the permissions to execute the desired operation.