Atomic Store: add `signup_is_store` site option to WP.com sites API endpoint
Summary:
Part of the Atomic Store signup flow work.

After a customer selects the "store" design type when signing up, this option is remembered in Calypso/Redux. Then, when a new site creation request is sent to WP.com, this option is passed as `options['designType'] = 'store'` to the site's `options` table.

In this patch, we are adding this site option to the response of the WP.com Simple sites API endpoint so we know that the site should be treated as a store site in Calypso/everywhere. This is especially useful for showing different checkout-thank-you page after purchasing a plan and/or knowing that a transfer is happening on the Woo store installation page.

Test Plan: Sandbox the public-api. Run Calypso locally. Go through the `/start/atomic-store` signup flow and select the "store" design type on the first step. Then, continue the flow until the site is created. After that, go to dev console and do the `/sites/<site>` query. See if the `options` field contains `signup_is_store` set to `true`.

Reviewers: jsnajdr, jsnjohnston

Reviewed By: jsnajdr

Differential Revision: https://[private link]

Merges r163585-wpcom.