AT: add new site option called `has_pending_automated_transfer` to all sites which have a transfer pending on the backend
Summary:
Part of the Atomic Store signup flow work.

Last week, we opened Atomic Store for internal testing. One of the bugs which were discovered is that if you navigate to Calypso main view before going to the `/store/<site>` page, you won't see the "Store" in the
 sidebar. Also, if the AT happening on the backend fails for some reason, you won't be notified in any way. What is more, even after the transfer finishes, Calypso won't be updated to reflect that (ie "Store" will
 still be missing in the sidebar).

We tried to solve this on the client-side by asking whether the site has `signup_is_store` site option and is younger than 10 minutes. However, this wouldn't work because if our customers don't finish the checkout
 step (don't buy the Business plan), we still keep the `signup_is_store` option but can't transfer.

In this patch, we have a different solution. We are adding a new site option called `has_pending_automated_transfer`. We will add this site option when starting the AT async job `at_atomic_store_initiate_transfer`
 so WP.com API clients (such as Calypso) know that a transfer is happening on the backend and can react accordingly.

In addition to that, we are removing this site option in `at_atomic_store_initiate_transfer` when we are sure the transfer has completed.

The only problem I see (lamosty) is that we don't remove this option if a transfer fails. There are dozens of places where a transfer can fail and it's almost impossible to catch them all (for example, if GT serve
r crashes mid-transfer, it wouldn't be possible to remove this option since GT handles the transfers). We will need to think about some time-based solution which would mark the transfer as failure after some time
since starting the transfer. We'll probably fix this on the client-side (Calypso) first as it's a blocker for Atomic Store.

Test Plan:
1. Boot up a local instance of Calypso on `atomic-store/handle-calypso-for-inprogress-transfers` branch - you'll be testing with that.
2. Apply this patch to your sandbox and sandbox the public-api.
3. Read https://fieldguide.automattic.com/asynchronous-jobs/testing-asynchronous-jobs-on-a-sandbox/ on how to test async jobs (ie rename `wpj_at_atomic_store_initiate_transfer` to something else and in `wpcom/wp-c
ontent/admin-plugins/wpcom-billing/products/bundles.php` add `error_log` to the `queue_async_job` so we'll know the job ID.
4. In local Calypso, run through the Atomic Store flow and purchase the Business plan. Stay on the checkout-thank-you page.
5. Open `/tmp/php-errors` on your sandbox: it should contain the job ID you queued in step 3.
6. Run the job with `php /home/wpcom/public_html/async-jobs/tester.php <job_id>`;
7. With dev console (https://[private link]) query the `/sites/<site>` endpoint and make sure the `options` object contains the `has_pending_automated_transfer` key set to `true`.
8. Wait till the transfer finishes (make sure it finishes successfully).
9. Repeat step 7 but make sure the `has_pending_automated_transfer` key is no longer there.

Reviewers: #team_stark, donpark, jsnajdr

Reviewed By: #team_stark, jsnajdr

Subscribers: jsnajdr, donpark, jsnjohnston

Differential Revision: https://[private link]

Merges r164091-wpcom.