Shortcodes: remove aws domain from Mailchimp shortcode. (#7989)
* Shortcodes: remove aws domain from Mailchimp shortcode.

Fixes #7945

To test, create a Maichimp form as explained here:
https://en.support.wordpress.com/mailchimp/

Then try to embed it into one of your posts or pages, when logged in as an editor. You should see the form transform into a shortcode on save. Make sure
the form is displayed properly on the site.

* Mailchimp: save a shortcode when pasting a form code in the widget

Fixes #7960
Fixes #7990

- The widget now calls the shortcode file even when the shortcodes are not active, so you do not need to activate the shortcode module for this to work.
- On save, the widget uses the reversal function provided in the shortcode to save a valid shortcode.
- That shortcode is then displayed in the widget, as before.

* Mailchimp shortcode: update failing unit test.