Changes include to include_once in shortcode files.
This is needed since #7991 has landed in Jetpack. There is a particular edge case that illustrates the problem. After #7991 the widget would check if a shortcode is present, and would load it if it wasn't. If then the shortcode would try to include that file again, it would fatal. This could happen when you have the widgets module enabled, and shortcode modules disabled. The fatal would appear in the request when you're trying to enable the shortcodes module.