Pr/7794 Redux -- Core 4.9 / Custom CSS (#8028)
* Add initial Jetpack CSS support for Code Editor control in core patch

* Comments detailing conditionals.

* We're conditionally adding this in the subsequent line.

Adding it in twice is bad -- the first time shouldn't need it?

May need to evaluate dependencies for what Core ships and bundles though for less/sass.

* Adjust CSS to handle height assignment.

This uses specificity to match Core styles and adjusts the height to accomodate core style changes.

* Remove unused return value variable.

* Roughing in new 4.9 version of the file.

We need to maintain an old version of the file for pre-4.9 versions of WordPress as we support vCurr and vCurr - 1.

* Tidy and purge.  Strip it down and then build it back up.

This should help clear out the cruft.  Let's start over and rearchitect the file.

* Errr, they're already jQuery objects by the documentation.

* Whitespace

* Override the `type` from `get_object_vars()`

We need this to be `jetpackCss`

* Simplify.  Initialize the core control to the right code format.

We don't really need to override it any longer, I don't think.

* Get the mode switching in place.

* Revert legacy file as we're working in new one.

* Run addLabels at document ready.

* Minor reorg.

* Whitespace tidy

* Create the div for the css links to go into.

* Whitespace tidying

* tabs > spaces

* Switch from watching the html dropdown to listening to events on the data model.

* Quote reserved word.

* CSS Tweaks for 4.9

* Activate linting after adding the gutters back in.

Otherwise the gutter may not get the errors displayed in it.

* Don't assume there are no other gutters.

Props @westonruter

* Remove the `wp` alias and clarify the `wp.customize` alias.

* Simplify and remove addDesc method.

* Remove toggling of the Gutters, Core handles this now.

As of https://core.trac.wordpress.org/changeset/41992

* Remove addTitle function and inline it.

Simpler to read code yields better maintainability.