Prevent multiple concurrent instances of the plugin upgrade routine
Adds a transient based lock with a short (10s) timeout to prevent
multiple processes from triggering the upgrade routine at the same time.

Only a single process after a version change should be responsible for
running the upgrade, to prevent excessive DB queries or other repeated
upgrade-related actions.

On many sites this is a non-issue, but on high traffic sites with many
concurrent requests, it can be a problem.

Fixes #8064