Open Graph: do not return images smaller than required size. (#6636)
* Open Graph: do not return images smaller than required size.

Fixes #2284

When using `jetpack_og_get_image`, we specify a minimum width and height (200px).
Those 2 values are then used by `Jetpack_PostImages::get_images()`. `get_images` looks for images following different methods, and we
applied the minimum width and height to most of them, thus excluding images that were too small. However, we did not apply the values for
the last fallback used to look for images in a post: `from_html`.

This commit implements the 2 width and height requirements to `from_html`. To do so, I switched from a regex match to using `DOMDocument`
to fetch the different image tags and their width and height values.

* Post Images: remove unnecessary isset.

@see https://github.com/Automattic/jetpack/pull/6636#discussion_r108501026

* Post Images: use stripos instead of stristr and add comment.

@see https://github.com/Automattic/jetpack/pull/6636#discussion_r108502010

* Post Images: cast width and height as int, & use empty over isset

@see https://github.com/Automattic/jetpack/pull/6636#discussion_r108502411

* Post Images: update test image string so it can be processed.

* Post Images: do not take into account if it is not set.

@see https://github.com/Automattic/jetpack/pull/6636/files#r108680308

* Post Images: fix unit tests to take into account the new size req.

* Post Images: avoid issues when parsing malformed HTML

@see #7919

* Post Images: create the $meta array as it was not created before.

@see https://github.com/Automattic/jetpack/pull/6636/files#r147718479

* Post Images: Suppress errors when dealing with malformed HTML.

@see #7919 and 256ca47dc42a24559f3e3c0441302a4f62a62e30