Jetpack: Adds WooCommerce bool for self-hosted w/Jetpack & WooCommerce enabled
Summary: This is a companion flag for is_wpcom_store to determine if the Jetpack site has WooCommerce enabled regardless if it's an Atomic site or not. This helps with the upcoming WooCommerce mobile apps to filter the results from me/sites to only show WooCommerce sites.

Test Plan: I'd like to add a couple unit tests if I can figure it out before this diff is ready for formal review. Otherwise, make sure woocommerce_is_active is false for all sites on .com, false for Jetpack sites without WooCommerce activated, and true for both wpcom Stores and Jetpack + WooCommerce active.

Reviewers: alexforcier, justin

Reviewed By: alexforcier, justin

Subscribers: justin

Differential Revision: https://[private link]

Merges r165439-wpcom.