Fix submit validation error when two instance of the widget are present in the same post.
https://github.com/Automattic/jetpack/issues/8154

There was a problem with jQuery selectors for source and target email inputs when the email sharing button is rendered more than one time in the same page.
The .val() method was always returning empty, causing an unexpected validation error.

The approach to fix the problem here is to use a selector relative to the current form element.