API: adding an `autosave` parameter to the posts/new endpoint
Will be true when a new post is initialy created as a draft on Calypso,
and will define `DOING_AUTOSAVE` to true.

This patch was created to clean out noise from the Activity Log and to
be more accurate / correct with defining `DOING_AUTOSAVE`

see: https://[private link]

Differential Revision: D8297

Merges r165562-wpcom.