CLI: Output site URL with some commands. (#8223)
Users running a multisite or scripting tools to run multiple single instances might be running Jetpack CLI commands across multiple sites at once. The lack of any information about what site the command is running on diminishes the usefulness of the output.

This change ensures the site URL is printed along with the output so that it's clear which site the command is running against.