Additional Filters For Images & Extended Delays
This PR proposes a couple of new filters for getting images most notably those within text-widgets. In addition, it also extends the delay to the PHP_MAX and the reason for this is because many plugins and themes i found are (and I am guilty of this too in the past) of putting their delays as 9999999 which albeit is a bad practice but because of this, many themes (notably from a particular online marketplace) aren't getting their images lazy loaded. 

This change allows more images to be affected by the filter and pushes the delay to the max that it can be without us just slapping a bunch of 9's on the delay.