Widget Visibility: fix widget minor conditions dropdown entities
The option items of the minor conditions dropdown in the widget
visibility settings are being correctly handled with esc_attr in the
backend. However, some of the category names may contain entities (e.g.
ampersands), so those option items must be decoded in the frontend to
make the text human-readable in these cases.