REST: Change /sites/:site/posts date range queries to use post_date_gmt
Summary: Currently we assume that that dates being passed to date range queries with /sites/:site/posts are in the same TZ as the site. That's pretty fragile assumption and can easily break depending on client preference. This patch changes how we build the query to always use UTC dates and post_date_gmt when querying, which allows clients to use any TZ they like in the query.

Test Plan:
- Install patch
- sandbox [private link]
- https://wordpress.com/read/blogs/71943556 should now page through all posts instead of getting stuck

Reviewers: bluefuton

Subscribers: westi

Differential Revision: D9104-code

This commit syncs r167670-wpcom.