Search: Set default filter count to 5 for all filters (#8453)
* Search: Set default filter count to 5 for all filters

* Search: Update JS default filter count fallback

Since we're passing the default filter count to Javascript via localization, and because there are plugins that move scripts all over the place to try and improve performance, I added a fallback in a previous commit. This fallback number shouldn't matter, but since this PR is to update the default to 5, we should also update it here.