Widgets: make update widget function place widgets in new sidebars
Previous commit in the patch removed it assuming that the sidebar already contains
the widget to be updated. This will not be the case in general, e.g. when placing
a widget in another sidebar. This commit brings it back conditional on the check
that the widget is absent from the sidebar.

In addition:
- it adds a helper function and leaves a couple of comments for clarity.
- small code refactor to avoid declaring unnecessary variables
- change error code when no sidebar is present
- updates error message when no widget is present in a sidebar.'