Work Around Trim() Parameter Error
Following this issue : https://github.com/Automattic/jetpack/issues/6563
One of the poster has isolated the bug to line 335 where trim() expects a string as an argument. I figured according to the function description, $instance is an array and therefore $value is also an array, in which trim() would not work properly. To work around this issue, I convert $value to an array for the sole purpose of this function. I have no way of compiling this myself but I hope this works.