Mailchimp Widget: avoid conflicts with jQuery UI (#8547)
* Fix issue #8546 - remove window.require and window.define from global scope after using a require definition

* Fix issue #8546 - instead of removing window.require and window.define from global scope and thus breaking mailchimp script, set window.define.amd to fool just real amd modules.

Mailchimp script seem to expect usable definitions of require and define, once we remove them it breaks. Since the purpose of this commit is to fix scripts like jquery-ui that will resort to define() if define.amd is defined - it is enough to unset define.amd. This is a very ugly solution, but it works.

* Update mailchimp shortcode test to match earlier fix for issue #8546 - test for updated script string.