Support design_type site option with the base class.
Summary:
Because the `design_type` option was supported in an endpoint under `me/sites`,
the option is not loaded when an API for a single site is requested.
This diff will move the `design_type` option support to the base site class,
so single site request can get responses that contain the option.

Test Plan:
1. Create a new website in Calypso
2. Move to the stats page.
3. Run `wpcom.req.get('/sites/YOUR_SITE_DOMAIN').then( res => console.log(res) );` in your browser console to log the response.
4. The response should include `design_type` option under the `options` field.

Reviewers: #marketing_developers

Subscribers: travwalter, fditrapani

Differential Revision: https://[private link]

Merges r168907-wpcom.