Expose the site's siteGoals option to the public api
siteGoals is an information that may be provided by users during signup, when they're asked what the primary goal of their site is.
For Google My Business (GMB), we will need that information back on the client to display a nudge to connect with GMB if the goal of the site matches some conditions (probably siteGoals === 'promote' for now).

Differential Revision: https://[private link]

Merges r169677-wpcom.