API: use a string with method_exists()
The method_exists() function can take an object or a string, but in this case
it was given what looks like a constant.  This logs a warning in PHP 7.2, and
in future versions of PHP it will be a fatal error.

I double checked with Ben Dwyer that this change looked ok.

#php7.2

Merges r170200-wpcom.