API: Jetpack: API: Updates /sites/%s/jetpack/modules endpoints response format to add override
Currently, it is possible to override the active modules in Jetpack and force some to be active or inactive. The issue with this is that our UI does not respond to that. So, while a module may be forced on, a user is shown UI that allows them to attempt to disable the module, except, when the page is refreshed, the module is still enabled.

See: https://github.com/Automattic/jetpack/pull/8836
See: https://github.com/Automattic/jetpack/issues/8800

Differential Revision: D10224-code

This commit syncs r170223-wpcom.