Jetpack Connect URL: use correct URL to auto-approve (#8881)
* `already_authorized` is a connection URL param that let's a connecting customer
know they've already connected the site.

`auth_approved` is a connection URL param that let's a connecting customer skip
the screen that shows the terms-of-service-acceptance button.

In #8544 we meant to use `auth_approved`

* Keep auth_approved param from WP Admin connect redir

This banners may include `auth_approved` which is intended to be sent to
the connection API endpoint. Ensure relevant `auth_approved` params are
included if provided when redirecting to the connection endpoint.