Translate Widget: default to responsive vertical layout. (#9022)
Fixes #5962 again.

When I worked on #7296 to first fix this issue, I wanted to change from the old unresponsive dropdown only layout (`2`)
to the new vertical layout that would be responsive (`0`).
But while addressing other feedback on the PR I reverted that change, ant the old layout remained the default.

This fixes that. The default is 0 (vertical layout) now.