Not showing connection view if the user cannot connect. (#9102)
* Not showing connection view if the user cannot connect.

Fixes #8066.

This PR adds a check to the connection view condition. Before we would not show anything to the user if they could not manage modules. Now we add the connect privilege check here as well, so the same thing happens if a user does not have the 'jetpack_connect' cap.

* Modified the privilege to check for connect cap instead of module management.