API: SAL: Remove skyword-tracking shortcode when querying for post via SAL (#9124)
This will remove the skyword-tracking shortcode, which is used on our main WordPress.com blog and Jetpack.com from posts that are returned via the API. Before this patch, the shortcode was being output unrendered. Even rendered, it would be a script tag which I'm not sure makes sense in content returned via the API.

This commit syncs r172151-wpcom.