Publicize: Change publicize form data format
Previously, publicize title/connections were passed from the editor
to the backend via $_POST (for Gutenberg and classic editor). This
allowed the back-end handling to be the same for both Gutenberg and
classic editor Publicize. A recent Gutenberg PR (#5971) changed
the api request format so that post data was passed in HTTP payload
instead of $_POST. Which broke the Publicize form in Gutenberg.
This commit registers a new 'publicize' REST field
for the 'post' endpoint. The update callback for the 'publicize'
field handles the Publicize logic directly instead of relying on
legacy $_POST processing that classic editor uses. The state handling
for PublicizeForm component is totally revamped in support of this
change. As part of the data handling improvement, the form state
is now stored only in editor only. withState and withDispatch are
utilized to prevent state data duplication within PublicizeForm component.

props @youknowriad