WooCommerce Analytics: always return a properly formatted URL.
When filtering the remove cart URL, we don't want to make any changes to the URL itself; we only want to add additional attributes if needed.

Fixes issues when the remove cart link was broken when the Product ID was missing from the cart item and Woo Analytics was adding it.