CLI: JPS: Change shebang back to sh in partner-provision.sh
Doing this to minimize the chance of errors due to refactoring. For example, if we switched to bash, and then used something that was support in bash but not sh, and if the host used sh to invoke the script, this would cause an error.

So, let's stick with sh for now.