Publicize: Move Publicize request processing to before publishing
Publicize request processing was on a register_rest_field callback.
This callback occurred right after a post is published, which is
too late because Publicize actions occur when post is published.
This resulted in missed Publicize settings if a post was not
lucky enough to be saved just before it was published.
This change moves the 'publicize' request field processing to
a 'rest_pre_insert_post' filter hook, which occurs right
before a post is saved/published.