Publicize: Add post ID to connection list request
Connection request Ajax endpoint didn't have the current $post
in context, so it couldn't properly check if the post was already
shared. This caused the Publicize form on the front end to not
be disabled after a post was already published. This commit
adds the post id to the 'wp_admin_get_publicize_connections'
Ajax endpoint to remedy this issue.