Code Sync: sync easy-markdown.php - was divergent (#9366)
* Easy Markdown: saving post_content_filtered even for autosaves

There is an issue with autosaves as they can't be properly restored as the post_content_filtered is not being populated for them.

If someone attempts to restore an autosave, they would get an empty texteditor, as the content is being read from post_content_filtered.

Populating and storing the post_content_filtered even for autosaves addresses this issue.

/z 50817

Merges r134728-wpcom.

* Markdown Bug Fixes and Updates. Also, Jetpack_Options plan sync.

This should bring four files into sync with Jetpack's versions. Hurrah!

@see https://[private link]

Merges r148841-wpcom.

* Markdown: Allow the `markdown` attribute on some HTML tags.

The Markdown Extra parser allows some tags to be marked with the `markdown` attribute, which means that Markdown inside that tag will be parsed.

Unfortunately, KSES strips the `markdown` attribute, so this fix tells KSES to allow it for supported tags.

Merges r156001-wpcom.

* Markdown: Tell TinyMCE not to strip the `markdown` attribute from HTML.

Following on from [156001], the editor needs to maintain the `markdown` attribute, rather than removing it while cleaning the HTML.

Merges r156056-wpcom.