Reuse markdown content on Meta Box update
The idea here is to reuse the existing original Markdown, on the second post for Meta Boxes following a `Save Draft` event. This way when the markdown parser runs a second time it does not parse the rendered markup in `post_content` and overwrite the original Markdown in `post_content_filtered`.