Publicize: Use get_filtered_connection_data method from separate branch
The connection data access method was moved to the separate
branch ('try/publicize-decouple-ui') This branch places the
connection helper methods in a different class, so the calls
to these methods needed to be updated. The 'disabled' field
for the connections also changed from a string to a boolean
so the frontend processing is updated as well.