Sharing: provide path instead of URL to fix errors loading js/css
Fixes #9470

`Jetpack::get_file_url_for_environment()` accepts paths, not full URLs,
but in #9238 we updated it to use full URLs for the sharing module.

That does not cause any issues in production environments, as the bug only appears when `SCRIPT_DEBUG` is set to true,
thus forcing `Jetpack::get_file_url_for_environment()` to use the full, unminified path.