Publicize: Adding close link to connection page
One concern with the flow for adding a connection
is the fact that a popup is opened and the user
may not realize this and my have a hard time
getting back to the editor. This experimental
change attempts to remedy this by putting a big
ugly banner on the popup setting page that, when
clicked, closes the pop up windows and bring
the user back to the editor. Style will
need some work if we decide to keep this
approach.