WPcom Merge: get site endpoint (#9474)
* API: get site endpoint

When doing a permission check, ensure we are always using the current blog's id.
For Jetpack sites, we'll need to grab it using get_current_blog_id instead of
using the given WPCOM blog id.

Differential Revision: D7311

Merges r163142-wpcom.

* Atomic Store: add `signup_is_store` site option to WP.com sites API endpoint

Summary:
Part of the Atomic Store signup flow work.

After a customer selects the "store" design type when signing up, this option is remembered in Calypso/Redux. Then, when a new site creation request is sent to WP.com, this option is passed as `options['designType'] = 'store'` to the site's `options` table.

In this patch, we are adding this site option to the response of the WP.com Simple sites API endpoint so we know that the site should be treated as a store site in Calypso/everywhere. This is especially useful for showing different checkout-thank-you page after purchasing a plan and/or knowing that a transfer is happening on the Woo store installation page.

Test Plan: Sandbox the public-api. Run Calypso locally. Go through the `/start/atomic-store` signup flow and select the "store" design type on the first step. Then, continue the flow until the site is created. After that, go to dev console and do the `/sites/<site>` query. See if the `options` field contains `signup_is_store` set to `true`.

Reviewers: jsnajdr, jsnjohnston

Reviewed By: jsnajdr

Differential Revision: https://[private link]

Merges r163585-wpcom.

* AT: add new site option called `has_pending_automated_transfer` to all sites which have a transfer pending on the backend

Summary:
Part of the Atomic Store signup flow work.

Last week, we opened Atomic Store for internal testing. One of the bugs which were discovered is that if you navigate to Calypso main view before going to the `/store/<site>` page, you won't see the "Store" in the
 sidebar. Also, if the AT happening on the backend fails for some reason, you won't be notified in any way. What is more, even after the transfer finishes, Calypso won't be updated to reflect that (ie "Store" will
 still be missing in the sidebar).

We tried to solve this on the client-side by asking whether the site has `signup_is_store` site option and is younger than 10 minutes. However, this wouldn't work because if our customers don't finish the checkout
 step (don't buy the Business plan), we still keep the `signup_is_store` option but can't transfer.

In this patch, we have a different solution. We are adding a new site option called `has_pending_automated_transfer`. We will add this site option when starting the AT async job `at_atomic_store_initiate_transfer`
 so WP.com API clients (such as Calypso) know that a transfer is happening on the backend and can react accordingly.

In addition to that, we are removing this site option in `at_atomic_store_initiate_transfer` when we are sure the transfer has completed.

The only problem I see (lamosty) is that we don't remove this option if a transfer fails. There are dozens of places where a transfer can fail and it's almost impossible to catch them all (for example, if GT serve
r crashes mid-transfer, it wouldn't be possible to remove this option since GT handles the transfers). We will need to think about some time-based solution which would mark the transfer as failure after some time
since starting the transfer. We'll probably fix this on the client-side (Calypso) first as it's a blocker for Atomic Store.

Test Plan:
1. Boot up a local instance of Calypso on `atomic-store/handle-calypso-for-inprogress-transfers` branch - you'll be testing with that.
2. Apply this patch to your sandbox and sandbox the public-api.
3. Read https://fieldguide.automattic.com/asynchronous-jobs/testing-asynchronous-jobs-on-a-sandbox/ on how to test async jobs (ie rename `wpj_at_atomic_store_initiate_transfer` to something else and in `wpcom/wp-c
ontent/admin-plugins/wpcom-billing/products/bundles.php` add `error_log` to the `queue_async_job` so we'll know the job ID.
4. In local Calypso, run through the Atomic Store flow and purchase the Business plan. Stay on the checkout-thank-you page.
5. Open `/tmp/php-errors` on your sandbox: it should contain the job ID you queued in step 3.
6. Run the job with `php /home/wpcom/public_html/async-jobs/tester.php <job_id>`;
7. With dev console (https://[private link]) query the `/sites/<site>` endpoint and make sure the `options` object contains the `has_pending_automated_transfer` key set to `true`.
8. Wait till the transfer finishes (make sure it finishes successfully).
9. Repeat step 7 but make sure the `has_pending_automated_transfer` key is no longer there.

Reviewers: #team_stark, donpark, jsnajdr

Reviewed By: #team_stark, jsnajdr

Subscribers: jsnajdr, donpark, jsnjohnston

Differential Revision: https://[private link]

Merges r164091-wpcom.

* Adds a `is_wpcom_store` flag to the options response, so we can easily figure out if a site is a Store on WP.com site or not. Also adds a few additional details to the Jetpack sync list.
Reviewed as part of https://[private link]

Merges r164771-wpcom.

* Jetpack: Adds WooCommerce bool for self-hosted w/Jetpack & WooCommerce enabled

Summary: This is a companion flag for is_wpcom_store to determine if the Jetpack site has WooCommerce enabled regardless if it's an Atomic site or not. This helps with the upcoming WooCommerce mobile apps to filter the results from me/sites to only show WooCommerce sites.

Test Plan: I'd like to add a couple unit tests if I can figure it out before this diff is ready for formal review. Otherwise, make sure woocommerce_is_active is false for all sites on .com, false for Jetpack sites without WooCommerce activated, and true for both wpcom Stores and Jetpack + WooCommerce active.

Reviewers: alexforcier, justin

Reviewed By: alexforcier, justin

Subscribers: justin

Differential Revision: https://[private link]

Merges r165439-wpcom.

* Support design_type site option with the base class.

Summary:
Because the `design_type` option was supported in an endpoint under `me/sites`,
the option is not loaded when an API for a single site is requested.
This diff will move the `design_type` option support to the base site class,
so single site request can get responses that contain the option.

Test Plan:
1. Create a new website in Calypso
2. Move to the stats page.
3. Run `wpcom.req.get('/sites/YOUR_SITE_DOMAIN').then( res => console.log(res) );` in your browser console to log the response.
4. The response should include `design_type` option under the `options` field.

Reviewers: #marketing_developers

Subscribers: travwalter, fditrapani

Differential Revision: https://[private link]

Merges r168907-wpcom.

* Expose the site's siteGoals option to the public api

siteGoals is an information that may be provided by users during signup, when they're asked what the primary goal of their site is.
For Google My Business (GMB), we will need that information back on the client to display a nudge to connect with GMB if the goal of the site matches some conditions (probably siteGoals === 'promote' for now).

Differential Revision: https://[private link]

Merges r169677-wpcom.

* REST API: Return wordads option when querying a Jetpack shadow site

Adds the `wordads` option to `$jetpack_response_option_additions` in the `WPCOM_JSON_API_GET_Site_Endpoint`, which will have the endpoint return the actual `wordads` state for a Jetpack site.

Differential Revision: https://[private link]

Merges r172358-wpcom.