EU Cookie Law Widget: Update the title and description (#9463)
This change proposes an update to the title and description of the EU
Cookie Law Widget. As it will soon be updated to also handle consent
(as required by the GDPR) on sites using WordAds (9373) — and we may
very well update it in the future to handle more privacy-related things
— it makes sense to expand the title/description accordingly.

To test, simply apply this change and make sure that the widget’s title
and description are updated accordingly. Also confirm that there are no
issues with adding/configuring/removing the widget. It should behave
exactly as before.