Privacy Settings: Content and Copy Updates (#9530)
* Privacy Settings: Language updates

This change updates the language contained within the Privacy settings.

* Privacy Settings: Add additional content and remove Sync doc tracking event

This change adds some additional content to the Privacy settings area
that points directly to the Jetpack Privacy Center. This also removes
the tracking event for the Jetpack Sync document link (as it has been
removed) and adds an event for the new Privacy Center link.

* Privacy Settings: Updated copy.

Minor copy updates.

* Privacy Settings: Fix some linting issues.

Fix linting issues.