WordAds: Update EU Cookie Law Widget w/ ads compliance features (#9373)
This change proposes an update to the title and description of the EU
Cookie Law Widget. As it will soon be updated to also handle consent
(as required by the GDPR) on sites using WordAds (9373) — and we may
very well update it in the future to handle more privacy-related things
— it makes sense to expand the title/description accordingly.

To test, simply apply this change and make sure that the widget’s title
and description are updated accordingly. Also confirm that there are no
issues with adding/configuring/removing the widget. It should behave
exactly as before.

* Update widget form for clarity

- Add the policy link pre-text to the default banner text section.
- Allow users to use line breaks in the banner text field.
- Reorder the sections in a more intuitive way.
- Update the labels for clarity.

* Combine adjacent translated strings