WordPress Posts Widget: Split out Base Class
WordPress.com and Jetpack do different things around data fetching and
storing. This way, both can inherit common code and implement their
specific requirements.

Dependency Injection might be a bit cleaner, but it wouldn't be a huge win
We'd have to have an encapsulating class named
Jetpack_Display_Posts_Widget for backward compatibility anyway.
We'd be able to get rid of the mocking we need for testing, but it's
already there, so… :)