Contact Form: Use associative keys for data exporters/erasers
In [#43931-core](https://core.trac.wordpress.org/ticket/43931) the decision was made to allow personal data exporters and erasers to be registered using associative key strings. This makes it easier to change what is in those lists of active callbacks because you don't have to iterate over the entire array to find what you're looking for.

This commit adds `jetpack-contact-form` as the key string for both the exporter and the eraser.