Contact Form: Return booleans in eraser callback
In `personal_data_eraser`, the returned array includes items for
`items_removed` and `items_retained`. In Core, these were originally
intended to contain an integer count of the number of items for each
outcome. However,
in [r43061](https://core.trac.wordpress.org/changeset/43061), this
was changed so they just contained booleans instead.

[#43602-core](https://core.trac.wordpress.org/ticket/43602) has some
conversation that kind of explains why this change was made, but it's
a bit hard to follow the thread.

This commit brings Grunion's `personal_data_eraser` method in line
with this change, by switching from integers to booleans.