Augment comments to explain interaction of microformat and excerpts.
When rendering excerpts, WP will strip tags, which makes the microformat
lat/long visible.  So, we need to abort microformat rendering in that
case.

When checking that we're in the excerpt filter, we need to grab the
wp_current_filter array from the global scope directly because WP core's
current_filter() function only looks at the last element in that array,
which in this context will be "the_content", not "get_the_excerpt".