Make test of sync queue lag much more reliable.
Previously, the test_queue_lag() method relied on sleeping for 3 seconds
after each call to the queue's add() method.  After sleeping for 6
seconds, it would then call lag() and assert that the intval() of the
return value was 6.  This would randomly fail on Travis builds when for
whatever reason the lag actually exceeded 7 seconds.

That may have been due to a slow database read or write.  The Jetpack
sync engine finds relevant options by doing a LIKE query on option_name,
which could have highly variable performance depending upon the number
of options in the table, other queries running at that time, etc.  So,
the lag goes above 6 seconds and the old test approach fails, which is
very frustrating.

Beyond that, the old approach required at least 6 seconds to complete
the test because of the sleep() calls.  Slow tests are generally
undesirable, obviously.

This commit instead makes the test more deterministic by mocking and
injecting the time values.  When adding items to the queue, the new
generate_option_name_timestamp() method is called.  During testing, we
mock this method and return predictable values.  Then in lag(), we allow
the caller to inject a value for $now, rather than being tightly coupled
to the PHP core microtime() function.  This allows us to also make that
call predictable and deterministic.

With these changes, we're still testing the lag() function to the same
extent we did previously.  The DB queries are still run and the same
logic is in place.  We just don't rely on sleep() and microtime()
directly during testing.