Simplify test values to make lag test easier to comprehend.
Because we're mocking the timestamp methods anyway, @enejb suggested we
simplify the values we're using in test_queue_lag().  This makes the
test easier to comprehend.

We do still need to use floating point numbers because the preg_match()
call in the lag() method is looking for `(\d+\.\d+)` when extracting the
timestamp from the option name.