Unify WordPress Posts Widget with WP.com's version (#9597)
* Move WordPress Posts Widget class to its own file

* WordPress Posts Widget: Split out Base Class

WordPress.com and Jetpack do different things around data fetching and
storing. This way, both can inherit common code and implement their
specific requirements.

Dependency Injection might be a bit cleaner, but it wouldn't be a huge win
We'd have to have an encapsulating class named
Jetpack_Display_Posts_Widget for backward compatibility anyway.
We'd be able to get rid of the mocking we need for testing, but it's
already there, so… :)

* WordPress Posts Widget: Document Back-Compat Reqs

* WordPress Posts Widget: Require Base Class

It helps to include the required files :)

* WordPress Posts Widget: Fix Tests

Update tests to match the code introduced in 69f5a302f45fed511a10ef49978889e6810dd783

* WordPress Posts Widget: Fix Style Path

* WordPress Posts Widgets: Return Newly Fetched Data

Return the newly fetched/processed data from `::update_instance()`.

* WordPress Posts Widget: Revert Fetch API Changes

Previously, the code to actually fetch the data from the WordPress.com
REST API was abstracted out so that WordPress.com could fetch the data
differently than Jetpack.

This turns out not to be necessary. Put the fetching code back in to the
base class.

* Revert "WordPress Posts Widget: Fix Tests"

This reverts commit a416e30dd47169710db886458ccce6421666b2cd.

These changes are no longer necessary now that the fetch API has been
restored to the original version.

* WordPress Posts Widget: Don't Trust the Details ;)